+++ CHANGELOG.rst
@@ -13,6 +13,10 @@ Added:
 
 - Ability to interpolate stimulator output waveforms from start timepoints of each subprotocol
 
+Fixed:
+^^^^^^
+- Calculation of twitch amplitude now uses Contraction 10% and Relaxation 90% points for baseline.
+
 
 0.28.3 (2022-12-08)
 -------------------